refactor: move pauli rotation tableau synthesis strategies to dedicated files
